+++ examples/interactive-map/src/App.tsx
@@ -208,7 +208,7 @@ const App: React.FC = () => {
               createCoordinates(2000, 1000),
             )}
           >
-            <Geographies geography="https://unpkg.com/world-atlas@2/countries-50m.json">
+            <Geographies geography="https://unpkg.com/world-atlas@2.0.2/countries-50m.json">
               {({ geographies }) =>
                 geographies.map((geo) => (
                   <Geography